LibWeb: Extract shared FFI plumbing from the Rust selector engine
Move RetainedCxxPointer, FfiCallScope, the borrowed DOM string view
types, and the UTF-16 comparison helpers into a new ffi_support module
so upcoming Rust ports can use the C++/Rust boundary toolkit without
reaching into the selector engine. Pure code motion; SelectorRustFFI.h
now also generates from the new module so the DomStringView layout
stays in the selector FFI namespace.
