+++ Libraries/LibWeb/CSS/Rust/src/lib.rs
@@ -15,6 +15,7 @@ pub mod css_pixels;
 mod css_tokenizer;
 pub mod custom_properties;
 pub mod ffi_stats;
+pub mod ffi_support;
 pub mod property_metadata;
 mod selector_engine;
 pub mod style_compute;
